requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.1+cpu
torchvision==0.6.1+cpu
aiogram==2.9.2
numpy~=1.18.1
Pillow~=7.1.2
matplotlib~=3.1.3